requirements.txt
pybullet
numpy
paho-mqtt